+++ src/layout/Layout.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import Header from './header';
 import Footer from './footer';
-import { useLanguage } from '../context/LanguajeContext';
+import { useLanguage } from '../hooks/useLanguage';
 
 const Layout = ({ children }) => {
